feat(eula): factor out a stored_path module
